迁移中英文对照到locales目录
