+++ src/locales/client/en.ts
@@ -1,0 +1,38 @@
+/** CLIENT LIST 字段（英文） */
+export const enClientTip = {
+  id: 'a unique 64-bit client ID',
+  addr: 'address/port of the client',
+  laddr: 'address/port of local address client connected to (bind address)',
+  fd: 'file descriptor corresponding to the socket',
+  name: 'the name set by the client with CLIENT SETNAME',
+  age: 'total duration of the connection in seconds',
+  idle: 'idle time of the connection in seconds',
+  flags: 'client flags (see below)',
+  db: 'current database ID',
+  sub: 'number of channel subscriptions',
+  psub: 'number of pattern matching subscriptions',
+  ssub: 'number of shard channel subscriptions. Added in Redis 7.0.3',
+  multi: 'number of commands in a MULTI/EXEC context',
+  watch: 'number of keys this client is currently watching. Added in Redis 7.4',
+  qbuf: 'query buffer length (0 means no query pending)',
+  qbufFree: 'free space of the query buffer (0 means the buffer is full)',
+  argvMem: 'incomplete arguments for the next command (already extracted from query buffer)',
+  multiMem: 'memory is used up by buffered multi commands. Added in Redis 7.0',
+  obl: 'output buffer length',
+  oll: 'output list length (replies are queued in this list when the buffer is full)',
+  omem: 'output buffer memory usage',
+  totMem: 'total memory consumed by this client in its various buffers',
+  events: 'file descriptor events (see below)',
+  cmd: 'last command played',
+  user: 'the authenticated username of the client',
+  redir: 'client id of current client tracking redirection',
+  resp: 'client RESP protocol version. Added in Redis 7.0',
+  rbp: "peak size of the client's read buffer since the client connected. Added in Redis 7.0",
+  rbs: "current size of the client's read buffer in bytes. Added in Redis 7.0",
+  libName: 'the name of the client library that is being used',
+  libVer: 'the version of the client library',
+  ioThread: 'id of I/O thread assigned to the client. Added in Redis 8.0',
+  totNetIn: 'total network input bytes read from this client',
+  totNetOut: 'total network output bytes sent to this client',
+  totCmds: 'total count of commands this client executed',
+} as const
