+++ src/locales/client/zh-cn.ts
@@ -1,0 +1,38 @@
+/** CLIENT LIST 字段（中文） */
+export const zhClientTip = {
+  id: '唯一的 64 位客户端 ID',
+  addr: '客户端的地址/端口',
+  laddr: '客户端连接到的本地地址/端口（绑定地址）',
+  fd: '对应于套接字的文件描述符',
+  name: '客户端使用 CLIENT SETNAME 设置的名称',
+  age: '连接的总持续时间（秒）',
+  idle: '连接的空闲时间（秒）',
+  flags: '客户端标志（见下文）',
+  db: '当前数据库 ID',
+  sub: '频道订阅数',
+  psub: '模式匹配订阅数',
+  ssub: '分片频道订阅数。在 Redis 7.0.3 中添加',
+  multi: 'MULTI/EXEC 上下文中的命令数',
+  watch: '此客户端当前正在监视的键数。在 Redis 7.4 中添加',
+  qbuf: '查询缓冲区长度（0 表示没有待处理的查询）',
+  qbufFree: '查询缓冲区的可用空间（0 表示缓冲区已满）',
+  argvMem: '下一个命令的不完整参数（已从查询缓冲区中提取）',
+  multiMem: '缓冲的多命令使用的内存。在 Redis 7.0 中添加',
+  obl: '输出缓冲区长度',
+  oll: '输出列表长度（当缓冲区满时，回复在此列表中排队）',
+  omem: '输出缓冲区内存使用情况',
+  totMem: '此客户端在其各种缓冲区中消耗的总内存',
+  events: '文件描述符事件（见下文）',
+  cmd: '执行的最后一条命令',
+  user: '客户端的已认证用户名',
+  redir: '当前客户端跟踪重定向的客户端 id',
+  resp: '客户端 RESP 协议版本。在 Redis 7.0 中添加',
+  rbp: '客户端连接以来其读取缓冲区的峰值大小。在 Redis 7.0 中添加',
+  rbs: '客户端读取缓冲区当前大小（字节）。在 Redis 7.0 中添加',
+  libName: '正在使用的客户端库的名称',
+  libVer: '客户端库的版本',
+  ioThread: '分配给客户端的 I/O 线程 ID。在 Redis 8.0 中添加',
+  totNetIn: '从该客户端读取的总网络输入字节数',
+  totNetOut: '发送给该客户端的总网络输出字节数',
+  totCmds: '该客户端执行的命令总数',
+} as const
