+++ src/locales/cmd/en.ts
@@ -1,20 +1,6 @@
-import { cloneDeep } from 'lodash'
-import { computed } from 'vue'
+import type { CommandHelpRow } from './types'
 
-import { isZh } from '@/utils/util'
-
-export interface CommandHelpRow {
-  key: string
-  title: string
-  group: string
-  summary: string
-  since: string
-  usage: string
-  description?: string
-}
-
-// 英文命令
-const enCommands: CommandHelpRow[] = [
+export const enCommands: CommandHelpRow[] = [
   {
     key: 'ACL CAT',
     title: 'ACL CAT',
@@ -4340,545 +4326,3 @@ const enCommands: CommandHelpRow[] = [
       'ZUNIONSTORE destination numkeys key [key ...] [WEIGHTS weight\n  [weight ...]] [AGGREGATE <SUM | MIN | MAX>]',
   },
 ]
-
-// 中文翻译 (千问)
-const zhCommands = cloneDeep(enCommands)
-const zhDescription = [
-  '列出ACL类别或类别中的命令',
-  '删除ACL用户并终止其连接',
-  '模拟用户执行命令但不实际执行',
-  '生成安全的伪随机密码用于标识ACL用户',
-  '列出用户的ACL规则',
-  '返回子命令的帮助文本',
-  '以ACL文件格式转储有效规则',
-  '从配置的ACL文件重新加载规则',
-  '列出由ACL规则触发的安全事件',
-  '将有效ACL规则保存到配置的ACL文件',
-  '创建和修改ACL用户及其规则',
-  '列出所有ACL用户',
-  '返回当前连接的认证用户名',
-  '访问控制列表命令的容器',
-  '将字符串追加到键的值中，若键不存在则创建',
-  '向集群客户端发送-ASK重定向信号',
-  '验证连接身份',
-  '将元素添加到布隆过滤器',
-  '返回布隆过滤器的基数',
-  '检查元素是否存在于布隆过滤器中',
-  '返回布隆过滤器的详细信息',
-  '添加一个或多个元素到布隆过滤器（若不存在则创建）',
-  '恢复通过SCANDUMP保存的过滤器',
-  '添加一个或多个元素到布隆过滤器（若不存在则创建）',
-  '检查一个或多个元素是否存在于布隆过滤器中',
-  '创建新的布隆过滤器',
-  '开始增量保存布隆过滤器',
-  '异步重写AOF文件到磁盘',
-  '异步保存数据库到磁盘',
-  '计算字符串中的位设置数（人口统计）',
-  '对字符串执行任意位域整数操作',
-  '对字符串执行只读位域整数操作',
-  '对多个字符串执行位运算并存储结果',
-  '查找字符串中第一个设置位（1）或清除位（0）',
-  '从列表弹出元素推送到另一列表并返回（阻塞直到可用）',
-  '从多个列表中弹出第一个元素（阻塞直到可用）',
-  '移除并返回列表的第一个元素（阻塞直到可用）',
-  '移除并返回列表的最后一个元素（阻塞直到可用）',
-  '从列表弹出元素推送到另一列表并返回（阻塞直到可用）',
-  '从一个或多个有序集合中按分数弹出成员（阻塞直到可用）',
-  '从一个或多个有序集合中弹出最高分成员（阻塞直到可用）',
-  '从一个或多个有序集合中弹出最低分成员（阻塞直到可用）',
-  '将元素添加到布谷鸟过滤器',
-  '若元素不存在则添加到布谷鸟过滤器',
-  '返回元素在布谷鸟过滤器中的存在次数',
-  '从布谷鸟过滤器中删除元素',
-  '检查一个或多个元素是否存在于布谷鸟过滤器',
-  '返回布谷鸟过滤器的详细信息',
-  '添加一个或多个元素到布谷鸟过滤器（若不存在则创建）',
-  '若元素不存在则添加到布谷鸟过滤器（若不存在则创建）',
-  '恢复通过SCANDUMP保存的过滤器',
-  '检查一个或多个元素是否存在于布谷鸟过滤器',
-  '创建新的布谷鸟过滤器',
-  '开始增量保存布隆过滤器',
-  '指示服务器是否跟踪下个请求的键',
-  '返回连接名称',
-  '返回连接的跟踪通知重定向客户端ID',
-  '返回子命令的帮助文本',
-  '返回连接的唯一客户端ID',
-  '返回连接的详细信息',
-  '终止开放连接',
-  '列出开放连接',
-  '设置连接的客户端驱逐模式',
-  '控制客户端命令是否影响键的LRU/LFU统计',
-  '暂停命令处理',
-  '指示服务器是否回复命令',
-  '设置客户端或连接特定信息',
-  '设置连接名称',
-  '控制服务器辅助客户端缓存',
-  '返回服务器辅助客户端缓存信息',
-  '从不同连接解除阻塞客户端',
-  '恢复暂停客户端的命令处理',
-  '客户端连接命令的容器',
-  '将新哈希槽分配给节点',
-  '将新哈希槽范围分配给节点',
-  '推进集群配置纪元',
-  '返回节点的活跃故障报告数',
-  '返回哈希槽中的键数量',
-  '将哈希槽设置为未绑定',
-  '将哈希槽范围设置为未绑定',
-  '强制副本节点执行手动故障转移',
-  '删除节点的所有槽信息',
-  '从节点表中移除节点',
-  '返回哈希槽中的键名',
-  '返回子命令帮助文本',
-  '返回节点的状态信息',
-  '返回键对应的哈希槽',
-  '返回所有与节点的TCP连接',
-  '强制节点与另一节点握手',
-  '管理原子槽迁移任务',
-  '返回节点ID',
-  '返回节点分片ID',
-  '返回节点的集群配置',
-  '列出主节点的副本节点',
-  '配置节点为另一节点的副本',
-  '重置节点',
-  '强制节点保存配置到磁盘',
-  '为新节点设置配置纪元',
-  '将哈希槽绑定到节点',
-  '返回集群槽到分片的映射',
-  '列出主节点的副本节点',
-  '返回当前节点分配的槽使用统计',
-  '返回集群槽到节点的映射',
-  'Redis集群命令的容器',
-  '增加一个或多个元素的计数器',
-  '返回草图的详细信息',
-  '按指定维度初始化计数-最小草图',
-  '按指定容错率初始化计数-最小草图',
-  '合并多个草图为一个草图',
-  '返回一个或多个元素在草图中的计数',
-  '返回命令数量统计',
-  '返回命令的文档信息',
-  '从任意命令中提取键名',
-  '从任意命令中提取键名和访问标志',
-  '返回子命令帮助文本',
-  '返回一个/多个/所有命令的信息',
-  '返回命令名称列表',
-  '返回所有命令的详细信息',
-  '返回配置参数的有效值',
-  '返回子命令帮助文本',
-  '重置服务器统计',
-  '持久化有效配置到文件',
-  '运行时设置配置参数',
-  '服务器配置命令的容器',
-  '复制键的值到新键',
-  '返回数据库中的键数量',
-  '调试命令的容器',
-  '将键的整数值减1（不存在则初始化为0）',
-  '将键的整数值减指定数（不存在则初始化为0）',
-  '删除一个或多个键',
-  '根据值或哈希摘要比较条件删除键',
-  '返回字符串值的哈希摘要（十六进制）',
-  '丢弃事务',
-  '返回键存储值的序列化表示',
-  '返回给定字符串',
-  '执行服务器端Lua脚本',
-  '通过SHA1摘要执行Lua脚本',
-  '通过SHA1摘要执行只读Lua脚本',
-  '执行只读服务器端Lua脚本',
-  '执行事务中的所有命令',
-  '检查一个或多个键是否存在',
-  '设置键的过期时间（秒）',
-  '设置键的过期时间（Unix时间戳）',
-  '返回键的过期时间（Unix时间戳）',
-  '启动服务器到副本的协调故障转移',
-  '调用函数',
-  '调用只读函数',
-  '删除所有数据库的所有键',
-  '删除当前数据库的所有键',
-  '在索引上执行搜索查询并进行聚合转换',
-  '添加索引别名',
-  '删除索引别名',
-  '添加或更新索引别名',
-  '向索引添加新字段',
-  '获取运行时配置选项',
-  '运行时配置选项帮助描述',
-  '设置运行时配置选项',
-  '按给定规格创建索引',
-  '删除游标',
-  '读取游标',
-  '向字典添加术语',
-  '从字典删除术语',
-  '转储字典中的所有术语',
-  '删除索引',
-  '返回复杂查询的执行计划',
-  '返回复杂查询的执行计划',
-  '执行结合文本搜索和向量相似度搜索的混合搜索',
-  '返回索引的信息和统计',
-  '执行FT.SEARCH或FT.AGGREGATE命令并收集性能信息',
-  '使用文本查询搜索索引，返回文档或ID',
-  '对查询执行拼写校正，返回拼写建议',
-  '向自动补全建议字典添加建议字符串',
-  '从建议索引删除字符串',
-  '获取前缀的自动补全建议',
-  '获取自动补全建议字典的大小',
-  '转储同义词组的内容',
-  '创建或更新包含附加术语的同义词组',
-  '返回Tag字段中的唯一标签',
-  '返回所有现有索引的列表',
-  '删除库及其函数',
-  '将所有库转储为二进制负载',
-  '删除所有库和函数',
-  '返回子命令帮助文本',
-  '终止执行中的函数',
-  '返回所有库的信息',
-  '创建库',
-  '从负载恢复所有库',
-  '返回执行中函数的信息',
-  '函数命令的容器',
-  '将成员添加到地理空间索引（若不存在则创建）',
-  '返回地理空间索引中两个成员的距离',
-  '返回地理空间索引成员的geohash字符串',
-  '返回地理空间索引成员的经纬度',
-  '查询地理空间索引中指定坐标范围内的成员',
-  '查询地理空间索引中指定成员范围内的成员',
-  '返回地理空间索引中指定成员范围内的成员',
-  '返回地理空间索引中指定坐标范围内的成员',
-  '查询地理空间索引中指定矩形或圆形区域内的成员',
-  '查询地理空间索引中指定区域的成员并存储结果',
-  '返回键的字符串值',
-  '返回指定偏移量的位值',
-  '返回删除键后的字符串值',
-  '返回设置过期时间后的字符串值',
-  '返回键存储的子字符串',
-  '返回设置新值前的旧字符串值',
-  '删除哈希中的一个或多个字段',
-  '与Redis服务器握手',
-  '检查字段是否存在于哈希中',
-  '使用相对时间（秒）设置哈希字段过期',
-  '使用绝对时间戳（秒）设置哈希字段过期',
-  '返回哈希字段的过期时间（秒）',
-  '返回哈希中字段的值',
-  '返回哈希中的所有字段和值',
-  '返回字段值并从哈希中删除',
-  '获取一个或多个哈希字段的值并设置过期',
-  '将哈希字段的整数值增加指定数（不存在则初始化为0）',
-  '将哈希字段的浮点数值增加指定数（不存在则初始化为0）',
-  '返回哈希中的所有字段',
-  '返回哈希中的字段数量',
-  '返回哈希中所有字段的值',
-  '设置多个字段的值',
-  '移除指定字段的过期时间',
-  '使用相对时间（毫秒）设置哈希字段过期',
-  '使用绝对时间戳（毫秒）设置哈希字段过期',
-  '返回哈希字段的过期时间（毫秒）',
-  '返回哈希字段的TTL（毫秒）',
-  '从哈希中随机返回一个或多个字段',
-  '迭代哈希的字段和值',
-  '创建或修改哈希字段的值',
-  '设置哈希字段的值并设置过期',
-  '仅当字段不存在时设置哈希字段的值',
-  '返回字段值的长度',
-  '返回哈希字段的TTL（秒）',
-  '返回哈希中的所有值',
-  '将键的整数值增加1（不存在则初始化为0）',
-  '将键的整数值增加指定数（不存在则初始化为0）',
-  '将键的浮点数值增加指定数（不存在则初始化为0）',
-  '返回服务器的信息和统计',
-  '将一个或多个JSON值追加到数组的末尾',
-  '返回JSON数组中第一个匹配值的索引',
-  '在数组的指定索引插入JSON值',
-  '返回JSON数组的长度',
-  '删除并返回数组指定索引的元素',
-  '修剪数组到指定索引范围',
-  '清空数组或对象的所有值，将数值设置为0',
-  '显示帮助信息',
-  '报告键的字节大小',
-  '调试容器命令',
-  '删除值',
-  '删除值',
-  '获取一个或多个路径的JSON序列化值',
-  '将JSON值合并到匹配路径',
-  '返回一个或多个键的路径值',
-  '设置或更新一个或多个键的JSON值',
-  '将路径的数值增加指定值',
-  '将路径的数值乘以指定值',
-  '返回JSON对象的键',
-  '返回JSON对象的键数量',
-  '返回JSON路径的RESP格式值',
-  '设置或更新JSON路径的值',
-  '将字符串追加到JSON字符串值',
-  '返回JSON字符串的长度',
-  '切换布尔值',
-  '返回JSON路径的类型',
-  '返回匹配模式的所有键名',
-  '返回上次成功保存到磁盘的Unix时间戳',
-  '返回人类可读的延迟分析报告',
-  '返回事件的延迟图',
-  '返回子命令帮助文本',
-  '返回命令延迟的累积分布',
-  '返回事件的时延样本',
-  '返回所有事件的最新延迟样本',
-  '重置一个或多个事件的延迟数据',
-  '延迟诊断命令的容器',
-  '查找最长公共子串',
-  '通过索引返回列表元素',
-  '在列表元素前后插入元素',
-  '返回列表长度',
-  '从列表弹出元素推送到另一列表并返回',
-  '从列表移除多个元素并返回',
-  '显示计算机艺术和Redis版本',
-  '从列表移除第一个元素并返回',
-  '返回列表中匹配元素的索引',
-  '在列表前添加一个或多个元素（不存在则创建）',
-  '仅当列表存在时在前添加元素',
-  '返回列表的指定范围元素',
-  '从列表中移除元素',
-  '通过索引设置列表元素的值',
-  '从列表两端移除元素',
-  '输出内存问题报告',
-  '返回子命令帮助文本',
-  '返回分配器统计信息',
-  '请求分配器释放内存',
-  '返回内存使用详情',
-  '估算键的内存使用',
-  '内存诊断命令的容器',
-  '原子获取一个或多个键的字符串值',
-  '将键从一个Redis实例原子转移到另一个',
-  '返回子命令帮助文本',
-  '返回所有已加载模块',
-  '加载模块',
-  '使用扩展参数加载模块',
-  '卸载模块',
-  '模块命令的容器',
-  '实时监听服务器接收的所有请求',
-  '将键移动到另一个数据库',
-  '原子创建或修改多个键的字符串值',
-  '原子设置多个键的共享过期时间',
-  '原子修改多个键的值（仅当键不存在时）',
-  '开始事务',
-  '返回Redis对象的内部编码',
-  '返回Redis对象的对数访问频率计数器',
-  '返回子命令帮助文本',
-  '返回Redis对象的上次访问时间',
-  '返回键值的引用计数',
-  '对象自省命令的容器',
-  '移除键的过期时间',
-  '设置键的过期时间（毫秒）',
-  '设置键的过期时间（Unix毫秒时间戳）',
-  '返回键的过期时间（Unix毫秒时间戳）',
-  '向HyperLogLog键添加元素（不存在则创建）',
-  '返回HyperLogLog键的基数估计',
-  'HyperLogLog值的内部调试命令',
-  '合并多个HyperLogLog值到一个键',
-  '用于测试HyperLogLog值的内部命令',
-  '返回服务器的活跃响应',
-  '设置键的字符串值和毫秒过期时间',
-  '监听匹配模式的频道消息',
-  '复制内部使用的命令',
-  '返回键的过期时间（毫秒）',
-  '向频道发布消息',
-  '返回活跃频道',
-  '返回子命令帮助文本',
-  '返回模式订阅数',
-  '返回频道订阅者数',
-  '返回活跃分片频道',
-  '返回分片频道订阅者数',
-  '发布/订阅命令的容器',
-  '停止监听匹配模式的频道消息',
-  '关闭连接',
-  '从数据库随机返回一个键名',
-  '启用集群副本节点的只读查询',
-  '启用集群副本节点的读写查询',
-  '重命名键并覆盖目标键',
-  '仅当目标键不存在时重命名',
-  '复制流配置内部命令',
-  '配置服务器为副本或提升为master',
-  '重置连接',
-  '集群迁移键的内部命令',
-  '从序列化值创建键',
-  '返回复制角色',
-  '返回并移除列表的最后元素',
-  '返回最后一个元素并推送到另一列表',
-  '向列表追加元素（不存在则创建）',
-  '仅当列表存在时追加元素',
-  '向集合添加成员（不存在则创建）',
-  '同步保存数据库到磁盘',
-  '迭代数据库中的键名',
-  '返回集合成员数量',
-  '设置服务器端Lua脚本的调试模式',
-  '检查服务器端Lua脚本是否存在于缓存',
-  '从缓存中移除所有服务器端Lua脚本',
-  '返回子命令帮助文本',
-  '终止执行中的服务器端Lua脚本',
-  '将服务器端Lua脚本加载到缓存',
-  'Lua脚本管理命令的容器',
-  '返回多个集合的差集',
-  '将多个集合的差集存储到键',
-  '更改所选数据库',
-  '设置键的字符串值忽略类型（不存在则创建）',
-  '设置或清除指定偏移量的位（不存在则创建）',
-  '设置键的字符串值和过期时间',
-  '仅当键不存在时设置字符串值',
-  '通过偏移量覆盖字符串部分内容',
-  '同步保存数据库到磁盘并关闭Redis服务器',
-  '返回多个集合的交集',
-  '返回多个集合交集的成员数',
-  '将多个集合的交集存储到键',
-  '判断成员是否属于集合',
-  '设置Redis服务器为副本或提升为master',
-  '返回慢查询日志条目',
-  '显示子命令帮助文本',
-  '返回慢查询日志条目数',
-  '清除所有慢查询日志条目',
-  '慢查询日志命令的容器',
-  '返回集合的所有成员',
-  '判断多个成员是否属于集合',
-  '将成员从一个集合移动到另一个',
-  '对列表、集合或有序集合进行排序',
-  '返回排序后的元素',
-  '从集合随机返回成员并删除',
-  '向分片频道发布消息',
-  '从集合中获取一个或多个随机成员',
-  '从集合中移除成员',
-  '迭代集合成员',
-  '监听分片频道的消息',
-  '返回字符串值的长度',
-  '监听频道的消息',
-  '返回字符串值的子串',
-  '返回多个集合的并集',
-  '将多个集合的并集存储到键',
-  '停止监听分片频道消息',
-  '交换两个Redis数据库',
-  '复制内部使用的命令',
-  '向t-digest草图添加观测值',
-  '返回每个输入排名的估计值',
-  '返回每个输入逆排名的估计值',
-  '返回每个输入值的分数估计',
-  '分配内存并初始化t-digest草图',
-  '返回t-digest草图的统计信息',
-  '返回t-digest草图的最大观测值',
-  '合并多个t-digest草图',
-  '返回t-digest草图的最小观测值',
-  '返回每个输入分数的估计值',
-  '返回每个输入值的排名估计',
-  '重置t-digest草图',
-  '返回每个输入值的逆排名估计',
-  '返回草图的均值估计',
-  '返回服务器时间',
-  '向Top-k草图添加项',
-  '返回一个或多个项的计数',
-  '增加一个或多个项的计数器',
-  '返回草图的详细信息',
-  '返回Top-k列表的完整项列表',
-  '检查一个或多个项是否存在于草图中',
-  '使用指定参数初始化TopK',
-  '返回指定键中存在键的数量',
-  '向时间序列追加样本',
-  '更新现有时间序列的保留策略',
-  '创建新的时间序列',
-  '创建压缩规则',
-  '减少最大时间戳样本的值',
-  '删除指定时间范围内的时间序列样本',
-  '删除压缩规则',
-  '获取时间序列中最大时间戳的样本',
-  '增加最大时间戳样本的值',
-  '返回时间序列的信息和统计',
-  '向一个或多个时间序列追加样本',
-  '获取每个匹配时间序列的最大时间戳样本',
-  '通过过滤器向前查询多个时间序列',
-  '通过过滤器向后查询多个时间序列',
-  '获取匹配过滤器的时间序列键',
-  '向前查询时间序列范围',
-  '向后查询时间序列范围',
-  '返回键的过期时间（秒）',
-  '确定键存储值的类型',
-  '异步删除一个或多个键',
-  '停止监听频道消息',
-  '忘记事务的监视键',
-  '向向量集中添加新元素或更新现有向量',
-  '返回向量集的元素数量',
-  '返回向量集的向量维度',
-  '返回元素关联的向量',
-  '检索元素的JSON属性',
-  '返回向量集的详细信息',
-  '检查元素是否存在于向量集中',
-  '返回元素在HNSW图各层的邻居',
-  '从向量集随机返回一个或多个成员',
-  '返回词法范围内的元素',
-  '从向量集中移除元素',
-  '关联或移除元素的JSON属性',
-  '通过向量相似度返回元素',
-  '阻塞直到连接的所有写命令完成复制',
-  '阻塞直到所有写命令写入AOF文件',
-  '监视键更改以确定事务执行',
-  '返回消费者组确认的消息数',
-  '确认并有条件删除流的消费者组条目',
-  '向流追加新消息（不存在则创建）',
-  '更改消费者组中消息的所有权',
-  '更改消费者组中单个消息的所有权',
-  '移除流中的消息并返回数量',
-  '从流中删除一个或多个条目',
-  '创建消费者组',
-  '创建消费者组中的消费者',
-  '从消费者组中删除消费者',
-  '销毁消费者组',
-  '返回子命令帮助文本',
-  '设置消费者组的最后交付ID',
-  '消费者组命令的容器',
-  '返回消费者组中的消费者列表',
-  '返回流的消费者组列表',
-  '返回子命令帮助文本',
-  '返回流的信息',
-  '流自省命令的容器',
-  '返回流中的消息数量',
-  '返回消费者组的待处理条目列表',
-  '返回流中指定ID范围的消息',
-  '从多个流获取大于指定ID的消息（阻塞直到可用）',
-  '为消费者组中的消费者获取新或历史消息（阻塞直到可用）',
-  '返回流中指定ID范围的逆序消息',
-  '复制流值的内部命令',
-  '从流开头删除消息',
-  '向有序集合添加成员（不存在则创建）',
-  '返回有序集合的成员数量',
-  '返回有序集合中指定分数范围的成员数',
-  '返回多个有序集合的差集',
-  '将多个有序集合的差集存储到键',
-  '增加有序集合成员的分数',
-  '返回多个有序集合的交集',
-  '返回交集的成员数量',
-  '将多个有序集合的交集存储到键',
-  '返回有序集合中词法范围的成员数',
-  '返回一个或多个有序集合的最高/最低分成员并删除',
-  '返回有序集合成员的分数',
-  '返回有序集合的最高分成员并删除',
-  '返回有序集合的最低分成员并删除',
-  '从有序集合随机返回一个或多个成员',
-  '返回有序集合中指定索引范围的成员',
-  '返回有序集合中词法范围的成员',
-  '返回有序集合中指定分数范围的成员',
-  '将有序集合的范围成员存储到键',
-  '返回有序集合中成员的升序索引',
-  '从有序集合中移除一个或多个成员',
-  '移除词法范围的成员',
-  '移除索引范围的成员',
-  '移除分数范围的成员',
-  '返回有序集合中指定索引范围的逆序成员',
-  '返回词法范围的逆序成员',
-  '返回分数范围的逆序成员',
-  '返回成员的降序索引',
-  '迭代有序集合的成员和分数',
-  '返回有序集合成员的分数',
-  '返回多个有序集合的并集',
-  '将多个有序集合的并集存储到键',
-]
-for (let i = 0; i < zhCommands.length; i++) {
-  zhCommands[i].summary = zhDescription[i]
-}
-
-// description: summary + since + group
-enCommands.forEach(command => {
-  command.description = `${command.summary} @since ${command.since} [${command.group}]`
-})
-zhCommands.forEach(command => {
-  command.description = `${command.summary} @since ${command.since} 【${command.group}】`
-})
-
-export const commandHelp = computed(() => (isZh.value ? zhCommands : enCommands))
