+++ src/locales/cmd/types.ts
@@ -1,0 +1,9 @@
+export interface CommandHelpRow {
+  key: string
+  title: string
+  group: string
+  summary: string
+  since: string
+  usage: string
+  description?: string
+}
