+++ src/locales/cmd/zh-cn.ts
@@ -1,0 +1,527 @@
+/** Redis 命令 help 中文摘要（仅 summary），key 与 `en.ts` 中 `CommandHelpRow.key` 一致。 */
+export const redisCmdSummaryZhCn: Record<string, string> = {
+  'ACL CAT': '列出ACL类别或类别中的命令',
+  'ACL DELUSER': '删除ACL用户并终止其连接',
+  'ACL DRYRUN': '模拟用户执行命令但不实际执行',
+  'ACL GENPASS': '生成安全的伪随机密码用于标识ACL用户',
+  'ACL GETUSER': '列出用户的ACL规则',
+  'ACL HELP': '返回子命令的帮助文本',
+  'ACL LIST': '以ACL文件格式转储有效规则',
+  'ACL LOAD': '从配置的ACL文件重新加载规则',
+  'ACL LOG': '列出由ACL规则触发的安全事件',
+  'ACL SAVE': '将有效ACL规则保存到配置的ACL文件',
+  'ACL SETUSER': '创建和修改ACL用户及其规则',
+  'ACL USERS': '列出所有ACL用户',
+  'ACL WHOAMI': '返回当前连接的认证用户名',
+  ACL: '访问控制列表命令的容器',
+  APPEND: '将字符串追加到键的值中，若键不存在则创建',
+  ASKING: '向集群客户端发送-ASK重定向信号',
+  AUTH: '验证连接身份',
+  'BF.ADD': '将元素添加到布隆过滤器',
+  'BF.CARD': '返回布隆过滤器的基数',
+  'BF.EXISTS': '检查元素是否存在于布隆过滤器中',
+  'BF.INFO': '返回布隆过滤器的详细信息',
+  'BF.INSERT': '添加一个或多个元素到布隆过滤器（若不存在则创建）',
+  'BF.LOADCHUNK': '恢复通过SCANDUMP保存的过滤器',
+  'BF.MADD': '添加一个或多个元素到布隆过滤器（若不存在则创建）',
+  'BF.MEXISTS': '检查一个或多个元素是否存在于布隆过滤器中',
+  'BF.RESERVE': '创建新的布隆过滤器',
+  'BF.SCANDUMP': '开始增量保存布隆过滤器',
+  BGREWRITEAOF: '异步重写AOF文件到磁盘',
+  BGSAVE: '异步保存数据库到磁盘',
+  BITCOUNT: '计算字符串中的位设置数（人口统计）',
+  BITFIELD: '对字符串执行任意位域整数操作',
+  BITFIELD_RO: '对字符串执行只读位域整数操作',
+  BITOP: '对多个字符串执行位运算并存储结果',
+  BITPOS: '查找字符串中第一个设置位（1）或清除位（0）',
+  BLMOVE: '从列表弹出元素推送到另一列表并返回（阻塞直到可用）',
+  BLMPOP: '从多个列表中弹出第一个元素（阻塞直到可用）',
+  BLPOP: '移除并返回列表的第一个元素（阻塞直到可用）',
+  BRPOP: '移除并返回列表的最后一个元素（阻塞直到可用）',
+  BRPOPLPUSH: '从列表弹出元素推送到另一列表并返回（阻塞直到可用）',
+  BZMPOP: '从一个或多个有序集合中按分数弹出成员（阻塞直到可用）',
+  BZPOPMAX: '从一个或多个有序集合中弹出最高分成员（阻塞直到可用）',
+  BZPOPMIN: '从一个或多个有序集合中弹出最低分成员（阻塞直到可用）',
+  'CF.ADD': '将元素添加到布谷鸟过滤器',
+  'CF.ADDNX': '若元素不存在则添加到布谷鸟过滤器',
+  'CF.COUNT': '返回元素在布谷鸟过滤器中的存在次数',
+  'CF.DEL': '从布谷鸟过滤器中删除元素',
+  'CF.EXISTS': '检查一个或多个元素是否存在于布谷鸟过滤器',
+  'CF.INFO': '返回布谷鸟过滤器的详细信息',
+  'CF.INSERT': '添加一个或多个元素到布谷鸟过滤器（若不存在则创建）',
+  'CF.INSERTNX': '若元素不存在则添加到布谷鸟过滤器（若不存在则创建）',
+  'CF.LOADCHUNK': '恢复通过SCANDUMP保存的过滤器',
+  'CF.MEXISTS': '检查一个或多个元素是否存在于布谷鸟过滤器',
+  'CF.RESERVE': '创建新的布谷鸟过滤器',
+  'CF.SCANDUMP': '开始增量保存布隆过滤器',
+  'CLIENT CACHING': '指示服务器是否跟踪下个请求的键',
+  'CLIENT GETNAME': '返回连接名称',
+  'CLIENT GETREDIR': '返回连接的跟踪通知重定向客户端ID',
+  'CLIENT HELP': '返回子命令的帮助文本',
+  'CLIENT ID': '返回连接的唯一客户端ID',
+  'CLIENT INFO': '返回连接的详细信息',
+  'CLIENT KILL': '终止开放连接',
+  'CLIENT LIST': '列出开放连接',
+  'CLIENT NO-EVICT': '设置连接的客户端驱逐模式',
+  'CLIENT NO-TOUCH': '控制客户端命令是否影响键的LRU/LFU统计',
+  'CLIENT PAUSE': '暂停命令处理',
+  'CLIENT REPLY': '指示服务器是否回复命令',
+  'CLIENT SETINFO': '设置客户端或连接特定信息',
+  'CLIENT SETNAME': '设置连接名称',
+  'CLIENT TRACKING': '控制服务器辅助客户端缓存',
+  'CLIENT TRACKINGINFO': '返回服务器辅助客户端缓存信息',
+  'CLIENT UNBLOCK': '从不同连接解除阻塞客户端',
+  'CLIENT UNPAUSE': '恢复暂停客户端的命令处理',
+  CLIENT: '客户端连接命令的容器',
+  'CLUSTER ADDSLOTS': '将新哈希槽分配给节点',
+  'CLUSTER ADDSLOTSRANGE': '将新哈希槽范围分配给节点',
+  'CLUSTER BUMPEPOCH': '推进集群配置纪元',
+  'CLUSTER COUNT-FAILURE-REPORTS': '返回节点的活跃故障报告数',
+  'CLUSTER COUNTKEYSINSLOT': '返回哈希槽中的键数量',
+  'CLUSTER DELSLOTS': '将哈希槽设置为未绑定',
+  'CLUSTER DELSLOTSRANGE': '将哈希槽范围设置为未绑定',
+  'CLUSTER FAILOVER': '强制副本节点执行手动故障转移',
+  'CLUSTER FLUSHSLOTS': '删除节点的所有槽信息',
+  'CLUSTER FORGET': '从节点表中移除节点',
+  'CLUSTER GETKEYSINSLOT': '返回哈希槽中的键名',
+  'CLUSTER HELP': '返回子命令帮助文本',
+  'CLUSTER INFO': '返回节点的状态信息',
+  'CLUSTER KEYSLOT': '返回键对应的哈希槽',
+  'CLUSTER LINKS': '返回所有与节点的TCP连接',
+  'CLUSTER MEET': '强制节点与另一节点握手',
+  'CLUSTER MIGRATION': '管理原子槽迁移任务',
+  'CLUSTER MYID': '返回节点ID',
+  'CLUSTER MYSHARDID': '返回节点分片ID',
+  'CLUSTER NODES': '返回节点的集群配置',
+  'CLUSTER REPLICAS': '列出主节点的副本节点',
+  'CLUSTER REPLICATE': '配置节点为另一节点的副本',
+  'CLUSTER RESET': '重置节点',
+  'CLUSTER SAVECONFIG': '强制节点保存配置到磁盘',
+  'CLUSTER SET-CONFIG-EPOCH': '为新节点设置配置纪元',
+  'CLUSTER SETSLOT': '将哈希槽绑定到节点',
+  'CLUSTER SHARDS': '返回集群槽到分片的映射',
+  'CLUSTER SLAVES': '列出主节点的副本节点',
+  'CLUSTER SLOT-STATS': '返回当前节点分配的槽使用统计',
+  'CLUSTER SLOTS': '返回集群槽到节点的映射',
+  CLUSTER: 'Redis集群命令的容器',
+  'CMS.INCRBY': '增加一个或多个元素的计数器',
+  'CMS.INFO': '返回草图的详细信息',
+  'CMS.INITBYDIM': '按指定维度初始化计数-最小草图',
+  'CMS.INITBYPROB': '按指定容错率初始化计数-最小草图',
+  'CMS.MERGE': '合并多个草图为一个草图',
+  'CMS.QUERY': '返回一个或多个元素在草图中的计数',
+  'COMMAND COUNT': '返回命令数量统计',
+  'COMMAND DOCS': '返回命令的文档信息',
+  'COMMAND GETKEYS': '从任意命令中提取键名',
+  'COMMAND GETKEYSANDFLAGS': '从任意命令中提取键名和访问标志',
+  'COMMAND HELP': '返回子命令帮助文本',
+  'COMMAND INFO': '返回一个/多个/所有命令的信息',
+  'COMMAND LIST': '返回命令名称列表',
+  COMMAND: '返回所有命令的详细信息',
+  'CONFIG GET': '返回配置参数的有效值',
+  'CONFIG HELP': '返回子命令帮助文本',
+  'CONFIG RESETSTAT': '重置服务器统计',
+  'CONFIG REWRITE': '持久化有效配置到文件',
+  'CONFIG SET': '运行时设置配置参数',
+  CONFIG: '服务器配置命令的容器',
+  COPY: '复制键的值到新键',
+  DBSIZE: '返回数据库中的键数量',
+  DEBUG: '调试命令的容器',
+  DECR: '将键的整数值减1（不存在则初始化为0）',
+  DECRBY: '将键的整数值减指定数（不存在则初始化为0）',
+  DEL: '删除一个或多个键',
+  DELEX: '根据值或哈希摘要比较条件删除键',
+  DIGEST: '返回字符串值的哈希摘要（十六进制）',
+  DISCARD: '丢弃事务',
+  DUMP: '返回键存储值的序列化表示',
+  ECHO: '返回给定字符串',
+  EVAL: '执行服务器端Lua脚本',
+  EVALSHA: '通过SHA1摘要执行Lua脚本',
+  EVALSHA_RO: '通过SHA1摘要执行只读Lua脚本',
+  EVAL_RO: '执行只读服务器端Lua脚本',
+  EXEC: '执行事务中的所有命令',
+  EXISTS: '检查一个或多个键是否存在',
+  EXPIRE: '设置键的过期时间（秒）',
+  EXPIREAT: '设置键的过期时间（Unix时间戳）',
+  EXPIRETIME: '返回键的过期时间（Unix时间戳）',
+  FAILOVER: '启动服务器到副本的协调故障转移',
+  FCALL: '调用函数',
+  FCALL_RO: '调用只读函数',
+  FLUSHALL: '删除所有数据库的所有键',
+  FLUSHDB: '删除当前数据库的所有键',
+  'FT.AGGREGATE': '在索引上执行搜索查询并进行聚合转换',
+  'FT.ALIASADD': '添加索引别名',
+  'FT.ALIASDEL': '删除索引别名',
+  'FT.ALIASUPDATE': '添加或更新索引别名',
+  'FT.ALTER': '向索引添加新字段',
+  'FT.CONFIG GET': '获取运行时配置选项',
+  'FT.CONFIG HELP': '运行时配置选项帮助描述',
+  'FT.CONFIG SET': '设置运行时配置选项',
+  'FT.CREATE': '按给定规格创建索引',
+  'FT.CURSOR DEL': '删除游标',
+  'FT.CURSOR READ': '读取游标',
+  'FT.DICTADD': '向字典添加术语',
+  'FT.DICTDEL': '从字典删除术语',
+  'FT.DICTDUMP': '转储字典中的所有术语',
+  'FT.DROPINDEX': '删除索引',
+  'FT.EXPLAIN': '返回复杂查询的执行计划',
+  'FT.EXPLAINCLI': '返回复杂查询的执行计划',
+  'FT.HYBRID': '执行结合文本搜索和向量相似度搜索的混合搜索',
+  'FT.INFO': '返回索引的信息和统计',
+  'FT.PROFILE': '执行FT.SEARCH或FT.AGGREGATE命令并收集性能信息',
+  'FT.SEARCH': '使用文本查询搜索索引，返回文档或ID',
+  'FT.SPELLCHECK': '对查询执行拼写校正，返回拼写建议',
+  'FT.SUGADD': '向自动补全建议字典添加建议字符串',
+  'FT.SUGDEL': '从建议索引删除字符串',
+  'FT.SUGGET': '获取前缀的自动补全建议',
+  'FT.SUGLEN': '获取自动补全建议字典的大小',
+  'FT.SYNDUMP': '转储同义词组的内容',
+  'FT.SYNUPDATE': '创建或更新包含附加术语的同义词组',
+  'FT.TAGVALS': '返回Tag字段中的唯一标签',
+  'FT._LIST': '返回所有现有索引的列表',
+  'FUNCTION DELETE': '删除库及其函数',
+  'FUNCTION DUMP': '将所有库转储为二进制负载',
+  'FUNCTION FLUSH': '删除所有库和函数',
+  'FUNCTION HELP': '返回子命令帮助文本',
+  'FUNCTION KILL': '终止执行中的函数',
+  'FUNCTION LIST': '返回所有库的信息',
+  'FUNCTION LOAD': '创建库',
+  'FUNCTION RESTORE': '从负载恢复所有库',
+  'FUNCTION STATS': '返回执行中函数的信息',
+  FUNCTION: '函数命令的容器',
+  GEOADD: '将成员添加到地理空间索引（若不存在则创建）',
+  GEODIST: '返回地理空间索引中两个成员的距离',
+  GEOHASH: '返回地理空间索引成员的geohash字符串',
+  GEOPOS: '返回地理空间索引成员的经纬度',
+  GEORADIUS: '查询地理空间索引中指定坐标范围内的成员',
+  GEORADIUSBYMEMBER: '查询地理空间索引中指定成员范围内的成员',
+  GEORADIUSBYMEMBER_RO: '返回地理空间索引中指定成员范围内的成员',
+  GEORADIUS_RO: '返回地理空间索引中指定坐标范围内的成员',
+  GEOSEARCH: '查询地理空间索引中指定矩形或圆形区域内的成员',
+  GEOSEARCHSTORE: '查询地理空间索引中指定区域的成员并存储结果',
+  GET: '返回键的字符串值',
+  GETBIT: '返回指定偏移量的位值',
+  GETDEL: '返回删除键后的字符串值',
+  GETEX: '返回设置过期时间后的字符串值',
+  GETRANGE: '返回键存储的子字符串',
+  GETSET: '返回设置新值前的旧字符串值',
+  HDEL: '删除哈希中的一个或多个字段',
+  HELLO: '与Redis服务器握手',
+  HEXISTS: '检查字段是否存在于哈希中',
+  HEXPIRE: '使用相对时间（秒）设置哈希字段过期',
+  HEXPIREAT: '使用绝对时间戳（秒）设置哈希字段过期',
+  HEXPIRETIME: '返回哈希字段的过期时间（秒）',
+  HGET: '返回哈希中字段的值',
+  HGETALL: '返回哈希中的所有字段和值',
+  HGETDEL: '返回字段值并从哈希中删除',
+  HGETEX: '获取一个或多个哈希字段的值并设置过期',
+  HINCRBY: '将哈希字段的整数值增加指定数（不存在则初始化为0）',
+  HINCRBYFLOAT: '将哈希字段的浮点数值增加指定数（不存在则初始化为0）',
+  HKEYS: '返回哈希中的所有字段',
+  HLEN: '返回哈希中的字段数量',
+  HMGET: '返回哈希中所有字段的值',
+  HMSET: '设置多个字段的值',
+  HPERSIST: '移除指定字段的过期时间',
+  HPEXPIRE: '使用相对时间（毫秒）设置哈希字段过期',
+  HPEXPIREAT: '使用绝对时间戳（毫秒）设置哈希字段过期',
+  HPEXPIRETIME: '返回哈希字段的过期时间（毫秒）',
+  HPTTL: '返回哈希字段的TTL（毫秒）',
+  HRANDFIELD: '从哈希中随机返回一个或多个字段',
+  HSCAN: '迭代哈希的字段和值',
+  HSET: '创建或修改哈希字段的值',
+  HSETEX: '设置哈希字段的值并设置过期',
+  HSETNX: '仅当字段不存在时设置哈希字段的值',
+  HSTRLEN: '返回字段值的长度',
+  HTTL: '返回哈希字段的TTL（秒）',
+  HVALS: '返回哈希中的所有值',
+  INCR: '将键的整数值增加1（不存在则初始化为0）',
+  INCRBY: '将键的整数值增加指定数（不存在则初始化为0）',
+  INCRBYFLOAT: '将键的浮点数值增加指定数（不存在则初始化为0）',
+  INFO: '返回服务器的信息和统计',
+  'JSON.ARRAPPEND': '将一个或多个JSON值追加到数组的末尾',
+  'JSON.ARRINDEX': '返回JSON数组中第一个匹配值的索引',
+  'JSON.ARRINSERT': '在数组的指定索引插入JSON值',
+  'JSON.ARRLEN': '返回JSON数组的长度',
+  'JSON.ARRPOP': '删除并返回数组指定索引的元素',
+  'JSON.ARRTRIM': '修剪数组到指定索引范围',
+  'JSON.CLEAR': '清空数组或对象的所有值，将数值设置为0',
+  'JSON.DEBUG HELP': '显示帮助信息',
+  'JSON.DEBUG MEMORY': '报告键的字节大小',
+  'JSON.DEBUG': '调试容器命令',
+  'JSON.DEL': '删除值',
+  'JSON.FORGET': '删除值',
+  'JSON.GET': '获取一个或多个路径的JSON序列化值',
+  'JSON.MERGE': '将JSON值合并到匹配路径',
+  'JSON.MGET': '返回一个或多个键的路径值',
+  'JSON.MSET': '设置或更新一个或多个键的JSON值',
+  'JSON.NUMINCRBY': '将路径的数值增加指定值',
+  'JSON.NUMMULTBY': '将路径的数值乘以指定值',
+  'JSON.OBJKEYS': '返回JSON对象的键',
+  'JSON.OBJLEN': '返回JSON对象的键数量',
+  'JSON.RESP': '返回JSON路径的RESP格式值',
+  'JSON.SET': '设置或更新JSON路径的值',
+  'JSON.STRAPPEND': '将字符串追加到JSON字符串值',
+  'JSON.STRLEN': '返回JSON字符串的长度',
+  'JSON.TOGGLE': '切换布尔值',
+  'JSON.TYPE': '返回JSON路径的类型',
+  KEYS: '返回匹配模式的所有键名',
+  LASTSAVE: '返回上次成功保存到磁盘的Unix时间戳',
+  'LATENCY DOCTOR': '返回人类可读的延迟分析报告',
+  'LATENCY GRAPH': '返回事件的延迟图',
+  'LATENCY HELP': '返回子命令帮助文本',
+  'LATENCY HISTOGRAM': '返回命令延迟的累积分布',
+  'LATENCY HISTORY': '返回事件的时延样本',
+  'LATENCY LATEST': '返回所有事件的最新延迟样本',
+  'LATENCY RESET': '重置一个或多个事件的延迟数据',
+  LATENCY: '延迟诊断命令的容器',
+  LCS: '查找最长公共子串',
+  LINDEX: '通过索引返回列表元素',
+  LINSERT: '在列表元素前后插入元素',
+  LLEN: '返回列表长度',
+  LMOVE: '从列表弹出元素推送到另一列表并返回',
+  LMPOP: '从列表移除多个元素并返回',
+  LOLWUT: '显示计算机艺术和Redis版本',
+  LPOP: '从列表移除第一个元素并返回',
+  LPOS: '返回列表中匹配元素的索引',
+  LPUSH: '在列表前添加一个或多个元素（不存在则创建）',
+  LPUSHX: '仅当列表存在时在前添加元素',
+  LRANGE: '返回列表的指定范围元素',
+  LREM: '从列表中移除元素',
+  LSET: '通过索引设置列表元素的值',
+  LTRIM: '从列表两端移除元素',
+  'MEMORY DOCTOR': '输出内存问题报告',
+  'MEMORY HELP': '返回子命令帮助文本',
+  'MEMORY MALLOC-STATS': '返回分配器统计信息',
+  'MEMORY PURGE': '请求分配器释放内存',
+  'MEMORY STATS': '返回内存使用详情',
+  'MEMORY USAGE': '估算键的内存使用',
+  MEMORY: '内存诊断命令的容器',
+  MGET: '原子获取一个或多个键的字符串值',
+  MIGRATE: '将键从一个Redis实例原子转移到另一个',
+  'MODULE HELP': '返回子命令帮助文本',
+  'MODULE LIST': '返回所有已加载模块',
+  'MODULE LOAD': '加载模块',
+  'MODULE LOADEX': '使用扩展参数加载模块',
+  'MODULE UNLOAD': '卸载模块',
+  MODULE: '模块命令的容器',
+  MONITOR: '实时监听服务器接收的所有请求',
+  MOVE: '将键移动到另一个数据库',
+  MSET: '原子创建或修改多个键的字符串值',
+  MSETEX: '原子设置多个键的共享过期时间',
+  MSETNX: '原子修改多个键的值（仅当键不存在时）',
+  MULTI: '开始事务',
+  'OBJECT ENCODING': '返回Redis对象的内部编码',
+  'OBJECT FREQ': '返回Redis对象的对数访问频率计数器',
+  'OBJECT HELP': '返回子命令帮助文本',
+  'OBJECT IDLETIME': '返回Redis对象的上次访问时间',
+  'OBJECT REFCOUNT': '返回键值的引用计数',
+  OBJECT: '对象自省命令的容器',
+  PERSIST: '移除键的过期时间',
+  PEXPIRE: '设置键的过期时间（毫秒）',
+  PEXPIREAT: '设置键的过期时间（Unix毫秒时间戳）',
+  PEXPIRETIME: '返回键的过期时间（Unix毫秒时间戳）',
+  PFADD: '向HyperLogLog键添加元素（不存在则创建）',
+  PFCOUNT: '返回HyperLogLog键的基数估计',
+  PFDEBUG: 'HyperLogLog值的内部调试命令',
+  PFMERGE: '合并多个HyperLogLog值到一个键',
+  PFSELFTEST: '用于测试HyperLogLog值的内部命令',
+  PING: '返回服务器的活跃响应',
+  PSETEX: '设置键的字符串值和毫秒过期时间',
+  PSUBSCRIBE: '监听匹配模式的频道消息',
+  PSYNC: '复制内部使用的命令',
+  PTTL: '返回键的过期时间（毫秒）',
+  PUBLISH: '向频道发布消息',
+  'PUBSUB CHANNELS': '返回活跃频道',
+  'PUBSUB HELP': '返回子命令帮助文本',
+  'PUBSUB NUMPAT': '返回模式订阅数',
+  'PUBSUB NUMSUB': '返回频道订阅者数',
+  'PUBSUB SHARDCHANNELS': '返回活跃分片频道',
+  'PUBSUB SHARDNUMSUB': '返回分片频道订阅者数',
+  PUBSUB: '发布/订阅命令的容器',
+  PUNSUBSCRIBE: '停止监听匹配模式的频道消息',
+  QUIT: '关闭连接',
+  RANDOMKEY: '从数据库随机返回一个键名',
+  READONLY: '启用集群副本节点的只读查询',
+  READWRITE: '启用集群副本节点的读写查询',
+  RENAME: '重命名键并覆盖目标键',
+  RENAMENX: '仅当目标键不存在时重命名',
+  REPLCONF: '复制流配置内部命令',
+  REPLICAOF: '配置服务器为副本或提升为master',
+  RESET: '重置连接',
+  'RESTORE-ASKING': '集群迁移键的内部命令',
+  RESTORE: '从序列化值创建键',
+  ROLE: '返回复制角色',
+  RPOP: '返回并移除列表的最后元素',
+  RPOPLPUSH: '返回最后一个元素并推送到另一列表',
+  RPUSH: '向列表追加元素（不存在则创建）',
+  RPUSHX: '仅当列表存在时追加元素',
+  SADD: '向集合添加成员（不存在则创建）',
+  SAVE: '同步保存数据库到磁盘',
+  SCAN: '迭代数据库中的键名',
+  SCARD: '返回集合成员数量',
+  'SCRIPT DEBUG': '设置服务器端Lua脚本的调试模式',
+  'SCRIPT EXISTS': '检查服务器端Lua脚本是否存在于缓存',
+  'SCRIPT FLUSH': '从缓存中移除所有服务器端Lua脚本',
+  'SCRIPT HELP': '返回子命令帮助文本',
+  'SCRIPT KILL': '终止执行中的服务器端Lua脚本',
+  'SCRIPT LOAD': '将服务器端Lua脚本加载到缓存',
+  SCRIPT: 'Lua脚本管理命令的容器',
+  SDIFF: '返回多个集合的差集',
+  SDIFFSTORE: '将多个集合的差集存储到键',
+  SELECT: '更改所选数据库',
+  SET: '设置键的字符串值忽略类型（不存在则创建）',
+  SETBIT: '设置或清除指定偏移量的位（不存在则创建）',
+  SETEX: '设置键的字符串值和过期时间',
+  SETNX: '仅当键不存在时设置字符串值',
+  SETRANGE: '通过偏移量覆盖字符串部分内容',
+  SHUTDOWN: '同步保存数据库到磁盘并关闭Redis服务器',
+  SINTER: '返回多个集合的交集',
+  SINTERCARD: '返回多个集合交集的成员数',
+  SINTERSTORE: '将多个集合的交集存储到键',
+  SISMEMBER: '判断成员是否属于集合',
+  SLAVEOF: '设置Redis服务器为副本或提升为master',
+  'SLOWLOG GET': '返回慢查询日志条目',
+  'SLOWLOG HELP': '显示子命令帮助文本',
+  'SLOWLOG LEN': '返回慢查询日志条目数',
+  'SLOWLOG RESET': '清除所有慢查询日志条目',
+  SLOWLOG: '慢查询日志命令的容器',
+  SMEMBERS: '返回集合的所有成员',
+  SMISMEMBER: '判断多个成员是否属于集合',
+  SMOVE: '将成员从一个集合移动到另一个',
+  SORT: '对列表、集合或有序集合进行排序',
+  SORT_RO: '返回排序后的元素',
+  SPOP: '从集合随机返回成员并删除',
+  SPUBLISH: '向分片频道发布消息',
+  SRANDMEMBER: '从集合中获取一个或多个随机成员',
+  SREM: '从集合中移除成员',
+  SSCAN: '迭代集合成员',
+  SSUBSCRIBE: '监听分片频道的消息',
+  STRLEN: '返回字符串值的长度',
+  SUBSCRIBE: '监听频道的消息',
+  SUBSTR: '返回字符串值的子串',
+  SUNION: '返回多个集合的并集',
+  SUNIONSTORE: '将多个集合的并集存储到键',
+  SUNSUBSCRIBE: '停止监听分片频道消息',
+  SWAPDB: '交换两个Redis数据库',
+  SYNC: '复制内部使用的命令',
+  'TDIGEST.ADD': '向t-digest草图添加观测值',
+  'TDIGEST.BYRANK': '返回每个输入排名的估计值',
+  'TDIGEST.BYREVRANK': '返回每个输入逆排名的估计值',
+  'TDIGEST.CDF': '返回每个输入值的分数估计',
+  'TDIGEST.CREATE': '分配内存并初始化t-digest草图',
+  'TDIGEST.INFO': '返回t-digest草图的统计信息',
+  'TDIGEST.MAX': '返回t-digest草图的最大观测值',
+  'TDIGEST.MERGE': '合并多个t-digest草图',
+  'TDIGEST.MIN': '返回t-digest草图的最小观测值',
+  'TDIGEST.QUANTILE': '返回每个输入分数的估计值',
+  'TDIGEST.RANK': '返回每个输入值的排名估计',
+  'TDIGEST.RESET': '重置t-digest草图',
+  'TDIGEST.REVRANK': '返回每个输入值的逆排名估计',
+  'TDIGEST.TRIMMED_MEAN': '返回草图的均值估计',
+  TIME: '返回服务器时间',
+  'TOPK.ADD': '向Top-k草图添加项',
+  'TOPK.COUNT': '返回一个或多个项的计数',
+  'TOPK.INCRBY': '增加一个或多个项的计数器',
+  'TOPK.INFO': '返回草图的详细信息',
+  'TOPK.LIST': '返回Top-k列表的完整项列表',
+  'TOPK.QUERY': '检查一个或多个项是否存在于草图中',
+  'TOPK.RESERVE': '使用指定参数初始化TopK',
+  TOUCH: '返回指定键中存在键的数量',
+  'TS.ADD': '向时间序列追加样本',
+  'TS.ALTER': '更新现有时间序列的保留策略',
+  'TS.CREATE': '创建新的时间序列',
+  'TS.CREATERULE': '创建压缩规则',
+  'TS.DECRBY': '减少最大时间戳样本的值',
+  'TS.DEL': '删除指定时间范围内的时间序列样本',
+  'TS.DELETERULE': '删除压缩规则',
+  'TS.GET': '获取时间序列中最大时间戳的样本',
+  'TS.INCRBY': '增加最大时间戳样本的值',
+  'TS.INFO': '返回时间序列的信息和统计',
+  'TS.MADD': '向一个或多个时间序列追加样本',
+  'TS.MGET': '获取每个匹配时间序列的最大时间戳样本',
+  'TS.MRANGE': '通过过滤器向前查询多个时间序列',
+  'TS.MREVRANGE': '通过过滤器向后查询多个时间序列',
+  'TS.QUERYINDEX': '获取匹配过滤器的时间序列键',
+  'TS.RANGE': '向前查询时间序列范围',
+  'TS.REVRANGE': '向后查询时间序列范围',
+  TTL: '返回键的过期时间（秒）',
+  TYPE: '确定键存储值的类型',
+  UNLINK: '异步删除一个或多个键',
+  UNSUBSCRIBE: '停止监听频道消息',
+  UNWATCH: '忘记事务的监视键',
+  VADD: '向向量集中添加新元素或更新现有向量',
+  VCARD: '返回向量集的元素数量',
+  VDIM: '返回向量集的向量维度',
+  VEMB: '返回元素关联的向量',
+  VGETATTR: '检索元素的JSON属性',
+  VINFO: '返回向量集的详细信息',
+  VISMEMBER: '检查元素是否存在于向量集中',
+  VLINKS: '返回元素在HNSW图各层的邻居',
+  VRANDMEMBER: '从向量集随机返回一个或多个成员',
+  VRANGE: '返回词法范围内的元素',
+  VREM: '从向量集中移除元素',
+  VSETATTR: '关联或移除元素的JSON属性',
+  VSIM: '通过向量相似度返回元素',
+  WAIT: '阻塞直到连接的所有写命令完成复制',
+  WAITAOF: '阻塞直到所有写命令写入AOF文件',
+  WATCH: '监视键更改以确定事务执行',
+  XACK: '返回消费者组确认的消息数',
+  XACKDEL: '确认并有条件删除流的消费者组条目',
+  XADD: '向流追加新消息（不存在则创建）',
+  XAUTOCLAIM: '更改消费者组中消息的所有权',
+  XCLAIM: '更改消费者组中单个消息的所有权',
+  XDEL: '移除流中的消息并返回数量',
+  XDELEX: '从流中删除一个或多个条目',
+  'XGROUP CREATE': '创建消费者组',
+  'XGROUP CREATECONSUMER': '创建消费者组中的消费者',
+  'XGROUP DELCONSUMER': '从消费者组中删除消费者',
+  'XGROUP DESTROY': '销毁消费者组',
+  'XGROUP HELP': '返回子命令帮助文本',
+  'XGROUP SETID': '设置消费者组的最后交付ID',
+  XGROUP: '消费者组命令的容器',
+  'XINFO CONSUMERS': '返回消费者组中的消费者列表',
+  'XINFO GROUPS': '返回流的消费者组列表',
+  'XINFO HELP': '返回子命令帮助文本',
+  'XINFO STREAM': '返回流的信息',
+  XINFO: '流自省命令的容器',
+  XLEN: '返回流中的消息数量',
+  XPENDING: '返回消费者组的待处理条目列表',
+  XRANGE: '返回流中指定ID范围的消息',
+  XREAD: '从多个流获取大于指定ID的消息（阻塞直到可用）',
+  XREADGROUP: '为消费者组中的消费者获取新或历史消息（阻塞直到可用）',
+  XREVRANGE: '返回流中指定ID范围的逆序消息',
+  XSETID: '复制流值的内部命令',
+  XTRIM: '从流开头删除消息',
+  ZADD: '向有序集合添加成员（不存在则创建）',
+  ZCARD: '返回有序集合的成员数量',
+  ZCOUNT: '返回有序集合中指定分数范围的成员数',
+  ZDIFF: '返回多个有序集合的差集',
+  ZDIFFSTORE: '将多个有序集合的差集存储到键',
+  ZINCRBY: '增加有序集合成员的分数',
+  ZINTER: '返回多个有序集合的交集',
+  ZINTERCARD: '返回交集的成员数量',
+  ZINTERSTORE: '将多个有序集合的交集存储到键',
+  ZLEXCOUNT: '返回有序集合中词法范围的成员数',
+  ZMPOP: '返回一个或多个有序集合的最高/最低分成员并删除',
+  ZMSCORE: '返回有序集合成员的分数',
+  ZPOPMAX: '返回有序集合的最高分成员并删除',
+  ZPOPMIN: '返回有序集合的最低分成员并删除',
+  ZRANDMEMBER: '从有序集合随机返回一个或多个成员',
+  ZRANGE: '返回有序集合中指定索引范围的成员',
+  ZRANGEBYLEX: '返回有序集合中词法范围的成员',
+  ZRANGEBYSCORE: '返回有序集合中指定分数范围的成员',
+  ZRANGESTORE: '将有序集合的范围成员存储到键',
+  ZRANK: '返回有序集合中成员的升序索引',
+  ZREM: '从有序集合中移除一个或多个成员',
+  ZREMRANGEBYLEX: '移除词法范围的成员',
+  ZREMRANGEBYRANK: '移除索引范围的成员',
+  ZREMRANGEBYSCORE: '移除分数范围的成员',
+  ZREVRANGE: '返回有序集合中指定索引范围的逆序成员',
+  ZREVRANGEBYLEX: '返回词法范围的逆序成员',
+  ZREVRANGEBYSCORE: '返回分数范围的逆序成员',
+  ZREVRANK: '返回成员的降序索引',
+  ZSCAN: '迭代有序集合的成员和分数',
+  ZSCORE: '返回有序集合成员的分数',
+  ZUNION: '返回多个有序集合的并集',
+  ZUNIONSTORE: '将多个有序集合的并集存储到键',
+}
