+++ src/locales/config/index.ts
@@ -1,0 +1,8 @@
+import { computed } from 'vue'
+
+import { isZh } from '@/utils/util'
+
+import { enConfigTip } from './en'
+import { zhConfigTip } from './zh-cn'
+
+export const configTip = computed(() => (isZh.value ? zhConfigTip : enConfigTip))
