+++ src/views/tab/RedisClient.vue
@@ -3,7 +3,7 @@ import { computed, inject, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import type { AppMainShare } from '@/bindings/me-interface'
-import { clientTip as tips } from '@/utils/tip'
+import { clientTip as tips } from '@/locales/client'
 import { meConfirm, meHumanSeconds, meCommands, meOk } from '@/utils/util'
 import NodeList from '@/views/ext/NodeList.vue'
